+++ test/ShapeCrawler.Tests.Unit/ParagraphTests.cs
@@ -347,7 +347,7 @@ public class ParagraphTests : SCTest
     
     [Test]
     [SlideShape("062_text_spacing.pptx", 1, "TextBox 4", 50)]
-    public void Spacing_BeforeSpacingPoints_returns_before_spacing_in_points(IShape shape, double expectedPoints)
+    public void Spacing_BeforeSpacingPoints_returns_before_spacing_in_points(IShape shape, decimal expectedPoints)
     {
         // Arrange
         var paragraph = shape.TextBox!.Paragraphs[0];
@@ -358,7 +358,7 @@ public class ParagraphTests : SCTest
     
     [Test]
     [SlideShape("062_text_spacing.pptx", 1, "TextBox 4", 50)]
-    public void Spacing_AfterSpacingPoints_returns_after_spacing_in_points(IShape shape, double expectedPoints)
+    public void Spacing_AfterSpacingPoints_returns_after_spacing_in_points(IShape shape, decimal expectedPoints)
     {
         // Arrange
         var paragraph = shape.TextBox!.Paragraphs[0];
